refactor: enforce uid based access control

+++ service/delete_driver.go
@@ -22,6 +22,10 @@ func (service *DriverServiceServer) DeleteDriver(ctx context.Context,
 
 	driverId := strings.Split(req.Msg.Name, "/")[1]
 
+	if driverId != req.Header().Get("Authorization") {
+		return nil, connect.NewError(connect.CodePermissionDenied, errors.New("permission denied"))
+	}
+
 	_, err := service.driverRepository.DeleteDriver(ctx, driverId)
 
 	if err != nil {
